# docs/api/openapi/control-center.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgainERP Control Center API",
    "description": "\nCentral control plane REST API for AgainERP operators and Edge Agents.\n\n**Operator API** — prefix `/api/v1/*` — requires `Authorization: Bearer {jwt}` unless noted.\n\n**Agent API** — prefix `/agent/v1/*` — requires agent bearer token (mTLS in production).\n\n**Webhooks** — prefix `/webhooks/v1/*` — provider-signed inbound events.\n\nHigh-risk actions (e.g. registration approval) require MFA step-up within a 5-minute window when MFA is enabled.\n",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/registrations/": {
      "get": {
        "tags": [
          "registrations"
        ],
        "summary": "List Registrations",
        "operationId": "list_registrations_api_v1_registrations__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {},
                  "title": "Response List Registrations Api V1 Registrations  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "registrations"
        ],
        "summary": "Create Registration",
        "operationId": "create_registration_api_v1_registrations__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Registration Api V1 Registrations  Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/registrations/{registration_id}": {
      "get": {
        "tags": [
          "registrations"
        ],
        "summary": "Get Registration",
        "operationId": "get_registration_api_v1_registrations__registration_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "registration_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Registration Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Registration Api V1 Registrations  Registration Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/registrations/{registration_id}/approve": {
      "post": {
        "tags": [
          "registrations"
        ],
        "summary": "Approve Registration",
        "operationId": "approve_registration_api_v1_registrations__registration_id__approve_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "registration_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Registration Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationApprove"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Approve Registration Api V1 Registrations  Registration Id  Approve Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/registrations/{registration_id}/reject": {
      "post": {
        "tags": [
          "registrations"
        ],
        "summary": "Reject Registration",
        "operationId": "reject_registration_api_v1_registrations__registration_id__reject_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "registration_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Registration Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationReject"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Reject Registration Api V1 Registrations  Registration Id  Reject Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RegistrationApprove": {
        "properties": {
          "operator_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operator Notes"
          },
          "db_host": {
            "type": "string",
            "title": "Db Host",
            "default": "localhost"
          },
          "db_port": {
            "type": "integer",
            "title": "Db Port",
            "default": 5432
          },
          "db_user": {
            "type": "string",
            "title": "Db User",
            "default": "postgres"
          },
          "db_password": {
            "type": "string",
            "title": "Db Password",
            "default": "changeme"
          }
        },
        "type": "object",
        "title": "RegistrationApprove"
      },
      "RegistrationCreate": {
        "properties": {
          "business_name": {
            "type": "string",
            "title": "Business Name"
          },
          "contact_name": {
            "type": "string",
            "title": "Contact Name"
          },
          "contact_email": {
            "type": "string",
            "format": "email",
            "title": "Contact Email"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "requested_plan": {
            "type": "string",
            "title": "Requested Plan",
            "default": "starter"
          },
          "requested_modules": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Requested Modules",
            "default": [
              "catalog",
              "orders",
              "customers",
              "inventory"
            ]
          },
          "wants_ai": {
            "type": "boolean",
            "title": "Wants Ai",
            "default": false
          },
          "industry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Industry"
          },
          "deployment_mode": {
            "type": "string",
            "title": "Deployment Mode",
            "default": "saas"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region",
            "default": "Asia/Dhaka"
          },
          "website": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Website"
          },
          "employee_count": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Employee Count"
          },
          "referral_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Referral Source"
          },
          "operator_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operator Notes"
          }
        },
        "type": "object",
        "required": [
          "business_name",
          "contact_name",
          "contact_email"
        ],
        "title": "RegistrationCreate"
      },
      "RegistrationReject": {
        "properties": {
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "reason"
        ],
        "title": "RegistrationReject"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}